requirements.txt

drf-extra-fields==3.7.0
amqp==5.2.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.0
black==23.11.0
boto3==1.35.19
botocore==1.35.19
celery==5.4.0
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
Django==4.2.7
django-auditlog==3.0.0
django-better-admin-arrayfield==1.4.2
django-cors-headers==4.3.0
django-crum==0.7.9
django-extensions==3.2.3
django-filter==23.3
django-import-export==3.3.3
django-js-asset==2.2.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.16.0
gunicorn==23.0.0
identify==2.6.1
idna==3.10
inflection==0.5.1
isort==5.12.0
jmespath==1.0.1
kombu==5.3.7
MarkupPy==1.14
mypy-extensions==1.0.0
nodeenv==1.9.1
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.1
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.3
pre-commit==3.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycodestyle==2.11.1
PyJWT==2.9.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2023.3.post1
PyYAML==6.0.2
requests==2.31.0
s3transfer==0.10.2
sendgrid==6.11.0
six==1.16.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
tablib==3.5.0
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
django-storages==1.14.2
